fix: カテゴリーの編集ができない問題を修正
